fix: circular dependencies (#7889)
Closes: #7880
