+++ packages/blocks/src/data-view-block/data-view-block.ts
@@ -3,7 +3,7 @@ import {
   DeleteIcon,
   MoreHorizontalIcon,
 } from '@blocksuite/affine-components/icons';
-import { RangeManager } from '@blocksuite/block-std';
+import { RANGE_SYNC_EXCLUDE_ATTR } from '@blocksuite/block-std';
 import { Slice } from '@blocksuite/store';
 import { computed } from '@lit-labs/preact-signals';
 import { css, nothing, unsafeCSS } from 'lit';
@@ -223,7 +223,7 @@ export class DataViewBlockComponent extends CaptionedBlockComponent<DataViewBloc
   override connectedCallback() {
     super.connectedCallback();
 
-    this.setAttribute(RangeManager.rangeSyncExcludeAttr, 'true');
+    this.setAttribute(RANGE_SYNC_EXCLUDE_ATTR, 'true');
   }
 
   override renderBlock() {
