+++ packages/blocks/src/root-block/edgeless/components/text/edgeless-shape-text-editor.ts
@@ -1,6 +1,6 @@
 import { ThemeObserver } from '@blocksuite/affine-shared/theme';
 import {
-  RangeManager,
+  RANGE_SYNC_EXCLUDE_ATTR,
   ShadowlessElement,
   WithDisposable,
 } from '@blocksuite/block-std';
@@ -187,7 +187,7 @@ export class EdgelessShapeTextEditor extends WithDisposable(ShadowlessElement) {
 
   override connectedCallback() {
     super.connectedCallback();
-    this.setAttribute(RangeManager.rangeSyncExcludeAttr, 'true');
+    this.setAttribute(RANGE_SYNC_EXCLUDE_ATTR, 'true');
   }
 
   override firstUpdated(): void {
