+++ packages/framework/block-std/src/range/index.ts
@@ -1,3 +1,4 @@
+export * from './consts.js';
 export * from './inline-range-provider.js';
 export * from './range-binding.js';
 export * from './range-manager.js';
