+++ packages/framework/block-std/src/view/element/index.ts
@@ -1,4 +1,5 @@
 export * from './block-component.js';
+export * from './consts.js';
 export * from './gfx-block-component.js';
 export * from './lit-host.js';
 export * from './shadowless-element.js';
