+++ packages/framework/block-std/src/view/element/widget-component.ts
@@ -7,13 +7,10 @@ import { LitElement } from 'lit';
 import type { EventName, UIEventHandler } from '../../event/index.js';
 import type { BlockStdScope } from '../../scope/index.js';
 import type { BlockService } from '../../service/index.js';
+import type { BlockComponent } from './block-component.js';
 
 import { WithDisposable } from '../utils/with-disposable.js';
-import {
-  type BlockComponent,
-  modelContext,
-  serviceContext,
-} from './block-component.js';
+import { modelContext, serviceContext } from './consts.js';
 import { docContext, stdContext } from './lit-host.js';
 
 export class WidgetComponent<
