+++ packages/framework/store/src/store/doc/consts.ts
@@ -1,0 +1,5 @@
+export enum BlockViewType {
+  Bypass = 'bypass',
+  Display = 'display',
+  Hidden = 'hidden',
+}
