+++ packages/framework/store/src/store/doc/index.ts
@@ -1,4 +1,5 @@
 export * from './block/index.js';
 export * from './block-collection.js';
+export * from './consts.js';
 export * from './doc.js';
 export * from './query.js';
